Include vector embeddings for cards
This works well enough.  I intend to make it possible to say "which
particular parts" of the card description should match.  Also make a
public API in the main module. It's not quite user ffriendly at present

Also this commit turns out to be huge because it combines a bunch of
previous commits which i couldn't easily make attomic.