Start on Deck handler
This will likely require some code rearrangement to make the naming
reasonable, but the benefits should start accruing soon enough

In any case, it does lots of naive stuff because it is easy to follow
and the sizes whould be small.

We assume that all sideboard and owners must be singletons.  This is true
today, not sure about tomorrow