Allow for loading all collectible art
For ost use cases this will be enough.

We attempt to be kind with respect to the API server.  I'm not sure
where the lines are/should be.

I also don't know if the shuffle makes things worse (in terms of
locality) or better in terms of filling out changes.